Feature/service 22

+++ Hot4/Hot4.ViewModel/PaymentModel.cs
@@ -4,9 +4,9 @@
     {
         public long PaymentId { get; set; }
         public long AccountId { get; set; }
-        public int PaymentTypeId { get; set; }
+        public byte PaymentTypeId { get; set; }
         public string PaymentType { get; set; }
-        public int PaymentSourceId { get; set; }
+        public byte PaymentSourceId { get; set; }
         public string PaymentSource { get; set; }
         public decimal Amount { get; set; }
         public DateTime PaymentDate { get; set; }
