+++ Hot4/Hot4.ViewModel/PaymentTypeModel.cs
@@ -2,7 +2,7 @@
 {
     public class PaymentTypeModel
     {
-        public int PaymentTypeId { get; set; }
+        public byte PaymentTypeId { get; set; }
         public string PaymentType { get; set; }
         public string PaymentTypeText { get; set; }
     }
